requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
anitya_schema==2.3.0
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
blessed==1.20.0
blinker==1.9.0
boto3==1.35.99
botocore==1.35.99
bpython==0.25
cachetools==5.5.1
celery==5.4.0
cement==2.10.14
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
config==0.5.1
constantly==23.10.4
crochet==2.1.1
cryptography==44.0.0
curtsies==0.4.2
cwcwidth==0.1.9
daphne==4.1.2
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
distro==1.9.0
Django==5.1.5
django-allauth==65.3.1
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-countries==7.6.1
django-extensions==3.2.3
django-filter==24.3
django-geojson==4.1.0
django-ipware==7.0.1
django-phonenumber-field==8.0.0
django-picklefield==3.2
django-redis==5.4.0
django-ses==4.3.2
django-storages==1.14.4
django-tawkto==0.5.1
django_dramatiq==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
dramatiq==1.17.1
executing==2.2.0
fedora-messaging-the-new-hotness-schema==1.2.1
fedora_messaging==3.7.0
fiona==1.10.1
flower==2.0.1
frozenlist==1.5.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
googlemaps==4.10.0
greenlet==3.1.1
gssapi==1.9.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huey==2.5.2
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
koji==1.35.1
kombu==5.4.2
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
numba==0.61.0
numpy==2.1.3
onfido-python==4.6.0
openai==1.60.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.10.1
pexpect==4.9.0
phonenumbers==8.13.53
pika==1.3.2
pillow==11.1.0
pipwin==0.5.2
platformdirs==4.3.6
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pyogrio==0.10.0
pyOpenSSL==25.0.0
PyPrind==2.11.3
pyproj==3.7.0
pySmartDL==1.3.4
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-ipware==3.0.0
pytz==2024.2
pyxdg==0.28
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-gssapi==1.3.0
retry==0.9.2
retrying==1.3.4
rpds-py==0.22.3
s3transfer==0.10.0
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
service-identity==24.2.0
setuptools==75.8.0
shapely==2.0.6
six==1.17.0
smile-id-core==3.0.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
stripe==11.4.1
supervisor==4.2.5
termcolor==2.5.0
the-new-hotness==1.3.0
threadpoolctl==3.5.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.4.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
wrapt==1.17.2
yarl==1.18.3
zope.interface==7.2
